+++ scripts/plots.py
@@ -11,41 +11,14 @@ Includes:
 
 import plotly.graph_objects as go
 import plotly.express as px
-from helper_funcs import get_flight_destinations_from_airport_on_day, get_distance_vs_arr_delay, create_flight_direction_mapping_table, compute_wind_impact
+from db_queries import get_flight_destinations_from_airport_on_day, get_distance_vs_arr_delay
+from geo_utils import create_flight_direction_mapping_table, compute_wind_impact
 from constants import NYC_AIRPORTS
 from plotly.subplots import make_subplots
 import pandas as pd
 import numpy as np
 import sqlite3 as sql
 from math import sin, radians
-
-def combine_two_figures_side_by_side(fig1, fig2):
-    """
-    Combine two Plotly figures side by side in a single figure using subplots.
-    """
-    # Create a subplot figure with 1 row, 2 columns, and minimal horizontal spacing
-    combined_fig = make_subplots(rows=1, cols=2, horizontal_spacing=0)
-
-    # Add all traces from fig1 to the first subplot
-    for trace in fig1.data:
-        combined_fig.add_trace(trace, row=1, col=1)
-
-    # Add all traces from fig2 to the second subplot
-    for trace in fig2.data:
-        combined_fig.add_trace(trace, row=1, col=2)
-
-    # Optionally copy over any layout settings from fig1 or fig2 if needed
-    # e.g., combined_fig.layout.annotations = fig1.layout.annotations
-
-    # Remove margins or set them to your preference
-    combined_fig.update_layout(
-        margin=dict(l=0, r=0, t=0, b=0),
-        width=900,   # You can set a custom width/height
-        height=400,
-        showlegend=True  # Or False, depending on your preference
-    )
-
-    return combined_fig
 
 def plot_route_map(conn, origin, destination):
     """
@@ -1016,9 +989,6 @@ def plot_wind_direction(direction, wind_speed=1):
 
     return fig
 
-import pandas as pd
-import plotly.express as px
-
 def plot_avg_wind_speed_for_route(conn, origin, destination):
     """
     Plots the average daily wind speed for a given route (origin → destination)
@@ -1063,6 +1033,3 @@ def plot_avg_wind_speed_for_route(conn, origin, destination):
     )
 
     return fig
-
-
-
